\name{rcdklibs}
\alias{rcdklibs}
\title{CDK Libraries}
\description{
  This package provides the CDK libraries and is required by the
  \code{rcdk} package. In general, this package will change
  infrequently, thus avoiding large downloads.

  This package has no user accessible methods
}
\examples{
library(rcdklibs)
}
\author{Rajarshi Guha (\email{rguha@indiana.edu})}
\keyword{programming}
